Support one Ethernet interface in server mode.
Currently, Ethernet only supports interfaces in client mode
(e.g., to connect to the Internet). Add minimal support to
Ethernet to support interfaces in server mode. This simple
implementation only works on the default interface, which is the
first interface that is detected on boot (even if the interface
is later removed).

This also provides (but does not yet unhide) a simple API for
Tethering to request that the next-plugged-in interface to be
placed into server mode.

Test: Enabling tethering with change on top
Bug: 130840861
Change-Id: I78f85bc36aaceb62ce274003a75ea99d0b6bc8b7